add to predict_system
