+++ tests/st/test_cls/example_small_img_01.png
